pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "qrcode>=8.2",
    "web3>=7.14.0",
    "werkzeug>=3.1.4",
]
